Fix usage of musl and libbacktrace USE flag in DEPEND variable in dmd.eclass.
The text was supposed to say that setting the musl USE flag
should require the libbacktrace USE flag to also be set. By
trying to do this in the DEPEND variable, it made portage
think the dmd package already needed to be installed to
install the dmd package, because the musl USE flag requires
that. What really should be done is to use REQUIRED_USE to
specify USE flags depending on other USE flags from the same
ebuild, and that is what this commit makes it do now.