Updates and removals.
This is an (incomplete) attempt to update to EAPI 8,
and to remove errors that occur when running `eix-update`
with this repository installed on a system.

Some packages that I don't use anymore were removed as part
of this operation, as removing them was the easiest way
to remove the corresponding error messages.

There are still some things that need work (ex: kexec-tools
is still used somewhere, and doesn't look easy to update to
EAPI 8 without doing some work, such as merging recent
::gentoo ebuilds into this tree).